[Chore #157829475] Setup PostgreSQL and test suite for version two
