+++ server/controllers/requestsController.js
@@ -1,0 +1,17 @@
+
+export default class Request {
+  static getRequests(req, res) {
+
+  }
+
+  static getRequest(req, res) {
+
+  }
+
+  static setRequest(req, res) {
+
+  }
+
+  static editRequest(req, res) {
+  }
+}
