+++ server/models/database.js
@@ -1,0 +1,7 @@
+// // import pg from 'pg';
+// const connectionString = process.env.DATABASE_URL || 'postgres://localhost:5432/todo';
+// const client = new pg.Client('');
+// client.connect();
+// const query = client.query(
+//   'CREATE TABLE items(id SERIAL PRIMARY KEY, text VARCHAR(40) not null, complete BOOLEAN)');
+// query.on('end', () => { client.end(); })
