+++ server/package.json
@@ -36,7 +36,8 @@
   "dependencies": {
     "eslint-watch": "^3.1.4",
     "jsonwebtoken": "^8.2.1",
-    "nodemon": "^1.17.4"
+    "nodemon": "^1.17.4",
+    "pg": "^7.4.3"
   },
   "nyc": {
     "check coverage": true,
